+++ Modules/Intent.Modules.Java.Spring.Data.Repositories/Templates/TemplateExtensions.cs
@@ -10,12 +10,12 @@ namespace Intent.Modules.Java.Spring.Data.Repositories.Templates
 {
     public static class TemplateExtensions
     {
-        public static string GetEntityRepositoryName<T>(this IntentTemplateBase<T> template) where T : Intent.Modelers.Domain.Api.ClassModel
+        public static string GetEntityRepositoryName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Domain.Api.ClassModel
         {
             return template.GetTypeName(EntityRepositoryTemplate.TemplateId, template.Model);
         }
 
-        public static string GetEntityRepositoryName(this IntentTemplateBase template, Intent.Modelers.Domain.Api.ClassModel model)
+        public static string GetEntityRepositoryName(this IIntentTemplate template, Intent.Modelers.Domain.Api.ClassModel model)
         {
             return template.GetTypeName(EntityRepositoryTemplate.TemplateId, model);
         }