+++ Tests/Standard.SpringBoot.TestApplication/src/main/java/com/Intent.Modules.Java.Tests/Standard.SpringBoot.TestApplication/Application.java
@@ -1,4 +1,7 @@
 package com.Intent.Modules.Java.Tests.Standard.SpringBoot.TestApplication;
+import com.Intent.Modules.Java.Tests.Standard.SpringBoot.TestApplication.intent.IntentIgnore;
+import com.Intent.Modules.Java.Tests.Standard.SpringBoot.TestApplication.intent.IntentManage;
+import com.Intent.Modules.Java.Tests.Standard.SpringBoot.TestApplication.intent.Mode;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;