+++ Tests/Standard.SpringBoot.TestApplication/src/main/java/com/Intent.Modules.Java.Tests/Standard.SpringBoot.TestApplication/data/projections/OrderQueryProjection.java
@@ -1,0 +1,10 @@
+package com.Intent.Modules.Java.Tests.Standard.SpringBoot.TestApplication.data.projections;
+
+import java.util.UUID;
+
+
+public interface OrderQueryProjection {
+    String getNumber();
+
+    UUID getId();
+}