+++ Modules/Intent.Modules.Java.Persistence.JPA.Queries/release-notes.md
@@ -1,3 +1,8 @@
+### Version 3.4.0
+
+- Updated `QueryProjection` template role from `Data` to `Data.QueryProjection` to make it more generically discoverable.
+- Updated the signature of `GetQueryProjectionName()` to be more generic.
+
 ### Version 3.3.7
 
 - Updated supported client version to [3.3.16, 5.0.0-a).