+++ Modules/Intent.Modules.Java.Spring.Data.Repositories/release-notes.md
@@ -1,3 +1,9 @@
+### Version 4.2.0
+
+- Fixed Repository Query generation so that Entity name is used in the HQL query instead of the Repository name.
+- Updated `EntityRepository` template role from `Data` to `Data.Repositories` to make it more generically discoverable.
+- Updated the signature of `GetEntityRepositoryName()` to be more generic.
+
 ### Version 4.1.2
 
 - Updated dependencies and supported client versions to prevent warnings when used with Intent Architect 4.x.